tests: updated tests
